updated to work with PyQt5
PyQt5 changes the location of some GUI elements, by splitting up QtGui into QtGui and QtWidgets. Many previous calls to QtGui needed to be changed to QtWidgets instead. A call to .getInteger also needed to be changed to .getInt. Finally, in order for this to work, attempts to import the older PyQt4 were removed, since these changes are not backwards compatible.